requirements.txt
uvicorn
fastapi
python-multipart
pillow
scikit-learn